+++ src/autopas/cells/SortedCellView.h
@@ -90,7 +90,7 @@ class SortedCellView : public ParticleCell<Particle> {
    * Get the number of all particles stored in this cell (owned, halo and dummy).
    * @return number of particles stored in this cell (owned, halo and dummy).
    */
-  unsigned long size() const override { return _particles.size(); }
+  size_t size() const override { return _particles.size(); }
 
   /**
    * @copydoc autopas::ParticleCell::getNumberOfParticles()
